Purchases: Move ValidationErrorList to purchases components
